requirements.txt
Flask>=2.2
SQLAlchemy>=1.4
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
alembic>=1.11
Flask-APScheduler>=1.12
Flask-Caching[redis]>=2.0.0
Flask-Cors>=3.0.10
python-dotenv>=1.0
pymysql>=1.0
requests>=2.0
tenacity>=8.0.0
joblib>=1.3.0
pandas>=2.0
xgboost>=1.7
scikit-learn>=1.2
pytest>=7.0
pytest>=7.0
 